search optimizations

+++ structr/structr-rest/src/main/java/org/structr/rest/resource/EntityResolverResource.java
@@ -44,7 +44,7 @@ public class EntityResolverResource extends SortableResource {
 	}
 
 	@Override
-	public List<? extends GraphObject> doGet(String sortKey, boolean sortDescending, long pageSize, long page) throws FrameworkException {
+	public List<? extends GraphObject> doGet(String sortKey, boolean sortDescending, int pageSize, int page) throws FrameworkException {
 		throw new IllegalMethodException();
 	}
