+++ structr/structr-rest/src/main/java/org/structr/rest/resource/NamedRelationResource.java
@@ -64,7 +64,7 @@ public class NamedRelationResource extends WrappingResource {
 	}
 
 	@Override
-	public List<? extends GraphObject> doGet(String sortKey, boolean sortDescending, long pageSize, long page) throws FrameworkException {
+	public List<? extends GraphObject> doGet(String sortKey, boolean sortDescending, int pageSize, int page) throws FrameworkException {
 
 		List<GraphObject> relationResults = new LinkedList<GraphObject>();
 		if(wrappedResource != null) {
