+++ structr/structr-rest/src/main/java/org/structr/rest/resource/RelationshipResource.java
@@ -61,7 +61,7 @@ public class RelationshipResource extends WrappingResource {
 	}
 
 	@Override
-	public List<? extends GraphObject> doGet(String sortKey, boolean sortDescending, long pageSize, long page) throws FrameworkException {
+	public List<? extends GraphObject> doGet(String sortKey, boolean sortDescending, int pageSize, int page) throws FrameworkException {
 
 		List<? extends GraphObject> results = wrappedResource.doGet(sortKey, sortDescending, pageSize, page);
 		if(results != null && !results.isEmpty()) {
