+++ structr/structr-rest/src/main/java/org/structr/rest/resource/ViewFilterResource.java
@@ -76,7 +76,7 @@ public class ViewFilterResource extends WrappingResource {
 	}
 
 	@Override
-	public List<? extends GraphObject> doGet(String sortKey, boolean sortDescending, long pageSize, long page) throws FrameworkException {
+	public List<? extends GraphObject> doGet(String sortKey, boolean sortDescending, int pageSize, int page) throws FrameworkException {
 
 		if (wrappedResource != null) {
